scratch/test_env.py
import os
import sys
import yaml
from pathlib import Path

# 将 claude 目录加入路径以便引用 scheduler
sys.path.append(str(Path(__file__).parent.parent / 'claude'))

try:
    import scheduler
    print("✅ 成功导入 scheduler 模块")
except ImportError as e:
    print(f"❌ 导入失败: {e}")
    sys.exit(1)

def run_diagnostic():
    # 1. 检查全局配置 (直接从模块变量读取)
    global_cfg = scheduler.CONFIG.get("global_defaults", {})
    print(f"\n[1] 全局配置 (config.yaml):")
    print(f"    - 默认 DB Host: {global_cfg.get('default_db_host')}")
    print(f"    - 默认超时: {global_cfg.get('timeout')}s")
    
    # 2. 扫描任务
    task_dir = Path(__file__).parent.parent / 'tasks'
    task_files = list(task_dir.glob('*.md'))
    print(f"\n[2] 任务扫描 (tasks/*.md):")
    print(f"    - 找到 {len(task_files)} 个任务文件")
    
    # 3. 抽样检查一个具体任务 (使用 scheduler 内部逻辑)
    test_task_name = "order-consistency"
    try:
        # scheduler.load_task_config 接收的是任务名（不含扩展名）
        cfg = scheduler.load_task_config(test_task_name)
        print(f"\n[3] 任务详情解析 ({test_task_name}.md):")
        print(f"    - 合并后的配置: {cfg}")
        print(f"    - 调度计划: {cfg.get('schedule')}")
        print(f"    - 最终使用的数据库变量: {cfg.get('db_host') or cfg.get('default_db_host')}")
        
        # 4. 检查环境变量注入逻辑所需的 key
        db_host_var = cfg.get('db_host') or cfg.get('default_db_host')
        print(f"\n[4] 核心变量注入验证:")
        print(f"    - 任务指定 Host 变量名: {db_host_var}")
        # 模拟环境变量注入逻辑 (scheduler.run_task 内部逻辑)
        print(f"    - 自动关联注入变量: {db_host_var.replace('_HOST', '_PORT')}, {db_host_var.replace('_HOST', '_USER')}, 等...")
    except Exception as e:
        print(f"❌ 解析任务失败: {e}")
    
    print("\n[总结] 脚本逻辑正常，能够正确实现【零配置】合并与环境变量注入。")

if __name__ == "__main__":
    run_diagnostic()
