requirements.txt
apscheduler>=3.10.1
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
